added NymRequest constructor and destructor, restructured nym code
