+++ zingo-rpc/src/nym.rs
@@ -1,3 +1,4 @@
 //! Backend Nym functionality.
 
+pub mod client;
 pub mod utils;
