+++ zingo-rpc/src/walletrpc.rs
@@ -1,3 +1,4 @@
 //! Walletside nym gRPC implementations.
 
 pub mod service;
+pub mod utils;
